backend/requirements.txt
# FastAPI Voice Chatbot - Production Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI/ML Core
torch==2.1.1
transformers==4.35.2
accelerate==0.24.1

# HTTP Client for OpenRouter
httpx==0.25.2

# Audio Processing
faster-whisper==0.10.0
gTTS==2.4.0
pydub==0.25.1

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
requests==2.31.0

# Language Detection
langdetect==1.0.9

# Development
pytest==7.4.3
black==23.11.0